requirements.txt
# Core dependencies
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.0

# PyTorch (install separately with: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118)
# torch==2.0.1
# torchvision==0.15.2
# torchaudio==2.0.2

# YOLO and computer vision models
ultralytics==8.0.211
yolov5==7.0.13

# Utilities
pyyaml==6.0.1
tqdm==4.66.1
matplotlib==3.7.3
scikit-learn==1.3.0

# Development and logging
tensorboard==2.14.0
wandb==0.15.12
python-dotenv==1.0.0

# Optional: For video processing
imageio==2.33.0
imageio-ffmpeg==0.4.9
